Remove unused discord bot functionality
